Split client UI into domain modules
