+++ src/app/types.ts
@@ -1,0 +1,4 @@
+export type Tab = "review" | "queued" | "working" | "done";
+export type Inspector = "new-feed" | "add-source" | null;
+export type AttentionScreen = "feed" | "workspace" | "learnings";
+export type WorkspaceTab = "feed" | "global";
